+++ spec/lib/fitting/cover/json_schema_spec.rb
@@ -9,8 +9,8 @@ RSpec.describe Fitting::Cover::JSONSchema do
     let(:original) { MultiJson.load(File.read('spec/fixtures/required/0.json')) }
     let(:first_combination) { MultiJson.load(File.read('spec/fixtures/required/1.json')) }
     let(:second_combination) { MultiJson.load(File.read('spec/fixtures/required/2.json')) }
-    let(:first_details) { %w[required captcha] }
-    let(:second_details) { %w[required code] }
+    let(:first_details) { %w[required required.captcha] }
+    let(:second_details) { %w[required required.code] }
 
     it 'returns combinations' do
       expect(subject.combi).to eq([[first_combination, first_details], [second_combination, second_details]])
@@ -20,8 +20,8 @@ RSpec.describe Fitting::Cover::JSONSchema do
       let(:original) { MultiJson.load(File.read('spec/fixtures/required/attachments_0.json')) }
       let(:first_combination) { MultiJson.load(File.read('spec/fixtures/required/attachments_1.json')) }
       let(:second_combination) { MultiJson.load(File.read('spec/fixtures/required/attachments_2.json')) }
-      let(:first_details) { %w[required properties.result.login] }
-      let(:second_details) { %w[required properties.result.password] }
+      let(:first_details) { %w[required properties.result.required.login] }
+      let(:second_details) { %w[required properties.result.required.password] }
 
       it 'returns combinations' do
         expect(subject.combi).to eq([[first_combination, first_details], [second_combination, second_details]])